SelectAllButtons and Controller Refactoring
Added select all buttons for checkbox form controls. Additionally, split the main controller into multiple controllers for each feature.
