+++ Musician's Companion/CapstoneMusiciansCompanion/src/main/resources/templates/challengeGenerator.html
@@ -7,28 +7,51 @@
 		<a href="/MusiciansCompanion">Back</a>
 		
 		<form th:action="@{/MusiciansCompanion/Challenge}" th:object="${ChallengeSettings}" method="post">
-			<label for="Genre">Genre: </label>
-			<input th:each="genre: ${genres}" type="checkbox" id="Genre" th:text="${genre.name}" th:field=${ChallengeSettings.genres} th:value=${genre.id}>
+			<div class="Genres">
+				<label for="Genre">Genre: </label>
+				<input th:each="genre: ${genres}" type="checkbox" class="Genre" id="Genre" th:text="${genre.name}" th:field=${ChallengeSettings.genres} th:value=${genre.id}>
+				<button type='button' class="SelectAll" id="GenreSelectAll" data-target-class="Genre" data-state="Off">Select All</button>
+			</div>
 			<br>
 			
-			<label for="Mode">Mode: </label>
-			<input th:each="mode: ${modes}" type="checkbox" id="Mode" th:text="${mode.name}" th:field="${ChallengeSettings.modes}" th:value=${mode.id}>
+			<div class="Modes">
+				<label for="Mode">Mode: </label>
+				<input th:each="mode: ${modes}" type="checkbox" class="Mode" id="Mode" th:text="${mode.name}" th:field="${ChallengeSettings.modes}" th:value=${mode.id}>
+				<button type='button' class="SelectAll" id="ModeSelectAll" data-target-class="Mode" data-state="Off">Select All</button>
+			</div>
 			<br>
 			
-			<label for="Mood">Mood/Emotion: </label>
-			<input th:each="mood: ${moods}" type="checkbox" id="Mood" th:text="${mood.name}" th:field="${ChallengeSettings.moods}" th:value="${mood.id}">
+			<div class="Moods">
+				<label for="Mood">Mood/Emotion: </label>
+				<input th:each="mood: ${moods}" type="checkbox" class="Mood" id="Mood" th:text="${mood.name}" th:field="${ChallengeSettings.moods}" th:value="${mood.id}">
+				<button type='button' class="SelectAll" id="MoodSelectAll" data-target-class="Mood" data-state="Off">Select All</button>
+			</div>
 			<br>
 			
-			<label for="Note">Key: </label>
-			<div th:each="note: ${notes}">
-				<input th:if="${note.isSharp}" type="checkbox" id="Note" th:text="${note.noteLetter} + '#'" th:field="${ChallengeSettings.homeNotes}" th:value="${note.id}">
-				<input th:if="${note.isFlat}" type="checkbox" id="Note" th:text="${note.noteLetter} + 'b'" th:field="${ChallengeSettings.homeNotes}" th:value="${note.id}">
-				<input th:unless="${note.isSharp || note.isFlat}" type="checkbox" id="Note" th:text="${note.noteLetter}" th:field="${ChallengeSettings.homeNotes}" th:value="${note.id}">
+			<div class=Notes>
+				<label for="Note">Key: </label>
+				<div th:each="note: ${notes}">
+					<input th:if="${note.isSharp}" type="checkbox" class="Note" id="Note" th:text="${note.noteLetter} + '#'" th:field="${ChallengeSettings.homeNotes}" th:value="${note.id}">
+					<input th:if="${note.isFlat}" type="checkbox" class="Note" id="Note" th:text="${note.noteLetter} + 'b'" th:field="${ChallengeSettings.homeNotes}" th:value="${note.id}">
+					<input th:unless="${note.isSharp || note.isFlat}" type="checkbox" class="Note" id="Note" th:text="${note.noteLetter}" th:field="${ChallengeSettings.homeNotes}" th:value="${note.id}">
+				</div>
+			
+				<button type='button' class="SelectAll" id="NoteSelectAll" data-target-class="Note" data-state="Off">Select All</button>
 			</div>
 			
-			<label for="timeSig">Time Signature: </label>
-			<input th:each="timeSig: ${timeSigs}" type="checkbox" id="timeSig" th:text="${timeSig.beatPerMeasure} + '/' + ${timeSig.beatNote}" th:field="${ChallengeSettings.timeSigs}" th:value="${timeSig.id}">
+			<br>
+			
+			<div class="TimeSigs">
+				<label for="timeSig">Time Signature: </label>
+				<input th:each="timeSig: ${timeSigs}" type="checkbox" class="TimeSig" id="timeSig" th:text="${timeSig.beatPerMeasure} + '/' + ${timeSig.beatNote}" th:field="${ChallengeSettings.timeSigs}" th:value="${timeSig.id}">
+				<button type='button' class="SelectAll" id="TimeSigSelectAll" data-target-class="TimeSig" data-state="Off">Select All</button>
+			</div>
+			
 			<button type="submit">Run</button>
 		</form>
+		
+		
+		<script th:src="@{/selectAll.js}"></script>
+		
 	</body>
 </html>
