+++ Musician's Companion/CapstoneMusiciansCompanion/src/main/resources/templates/musiciansCompanion.html
@@ -8,6 +8,7 @@
 			<li><a href="/MusiciansCompanion/ChallengeGenerator">Challenge Generator</a></li>
 			<li><a href="/MusiciansCompanion/ChordChallengeGenerator">Chord Challenge</a></li>
 			<li><a href="/MusiciansCompanion/Metronome">Metronome</a></li>
+			<li><a href="/MusiciansCompanion/FindChordVoicings">Find Chord Voicings for Guitar</a></li>
 		</ul>
 		
 		<h2>Coming Soon</h2>
@@ -15,7 +16,7 @@
 			<li><a href="/MusiciansCompanion/Looper">Looper</a></li>
 			<li><a href="/MusiciansCompanion/Notation">Tabbing/Music Notation</a></li>
 			<li><a href="/MusiciansCompanion/AudioToRhythm">Convert Audio to Rhythm</a></li>
-			<li><a href="/MusiciansCompanion/FindChordVoicings">Find Chord Voicings for Guitar/Piano</a></li>
+			
 			<li><a href="/MusiciansCompanion/DrumMachine">DrumMachine</a></li>
 		</ul>
 	</body>
